requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
openai==0.28.0
packaging==24.1
pip>=23.0  # Update to the latest
propcache==0.2.0
python-dotenv==1.0.1
requests==2.32.3
setuptools>=67.0  # Update to the latest
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.1
yarl==1.17.1
zipp==3.20.2
